Added an intermediate base window object to handle file IO/communication
